Added app.yaml to deploy app to Google Cloud. Tested all endpoints and debugged issues. Updated tables in database to fix compatibility issues between models and tables.

+++ project/src/main/java/com/example/javadb/model/UserCommunityId.java
@@ -4,12 +4,12 @@ import java.io.Serializable;
 import java.util.Objects;
 
 public class UserCommunityId implements Serializable {
-    private Long user;  // Assuming User has a Long id
-    private Long community;  // Assuming CommunityGroup has a Long id
+    private Integer user;  // Assuming User has a Long id
+    private Integer community;  // Assuming CommunityGroup has a Long id
 
     public UserCommunityId() {}
 
-    public UserCommunityId(Long user, Long community) {
+    public UserCommunityId(Integer user, Integer community) {
         this.user = user;
         this.community = community;
     }
